+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 ==================
 
 ## 0.18.22
+### is.codion.common
+- PropertyStore.systemProperties() now uses a DefaultSystemPropertyFormatter, which formats class and module paths as one item per line.
 ### is.codion.common.rmi
 - ConnectionRequest.clientTypeId() renamed clientType().
 ### is.codion.common.model